Trap: Add working trap handler
The Core Local Interrupt Map (CLIM) is memory mapped at 0x0200_0000 for the
HiFive board, so we should be using write_volatile and read_volatile to
access the mip.

This is tested in lib.rs by triggering a software interrupt and
making sure that the assembly label trap_handler is called, and that
successfully jumps to the rust code in src/trap/mod.rs which clears the
pending interrupt in the CLIM and returns the mepc that was passed in to
the function from assembly. The assembly then resets the previous return
address to return to after the trap has been handled.